+++ crates/prelude/CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Patch
 
+- Build docs.rs for wasm32-unknown-unknown
+- Fix broken links in documentation
 - Update dependencies
 
 ## 0.4.0